Struct aws_sdk_cloudwatchlogs::client::fluent_builders::CreateExportTask

pub struct CreateExportTask<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateExportTask<C>

pub async fn send(
self
) -> Result<CreateExportTaskOutput, SdkError<CreateExportTaskError>> where
C: SmithyConnector,

pub fn task_name(self, input: impl Into<String>) -> Self

The name of the export task.

pub fn set_task_name(self, input: Option<String>) -> Self

pub fn log_group_name(self, input: impl Into<String>) -> Self

The name of the log group.

pub fn set_log_group_name(self, input: Option<String>) -> Self

pub fn log_stream_name_prefix(self, input: impl Into<String>) -> Self

Export only log streams that match the provided prefix. If you don't specify a value, no prefix filter is applied.

pub fn set_log_stream_name_prefix(self, input: Option<String>) -> Self

pub fn from(self, input: i64) -> Self

The start time of the range for the request, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp earlier than this time are not exported.

pub fn set_from(self, input: Option<i64>) -> Self

pub fn to(self, input: i64) -> Self

The end time of the range for the request, expreswatchlogsdocused as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not exported.

pub fn set_to(self, input: Option<i64>) -> Self

pub fn destination(self, input: impl Into<String>) -> Self

The name of S3 bucket for the exported log data. The bucket must be in the same AWS region.

pub fn set_destination(self, input: Option<String>) -> Self

pub fn destination_prefix(self, input: impl Into<String>) -> Self

The prefix used as the start of the key for every object exported. If you don't specify a value, the default is exportedlogs.

pub fn set_destination_prefix(self, input: Option<String>) -> Self