Module aws_sdk_cloudwatchlogs::error

Modules

data_already_accepted_exception

See DataAlreadyAcceptedException

invalid_operation_exception

See InvalidOperationException

invalid_parameter_exception

See InvalidParameterException

invalid_sequence_token_exception

See InvalidSequenceTokenException

limit_exceeded_exception

See LimitExceededException

malformed_query_exception

See MalformedQueryException

operation_aborted_exception

See OperationAbortedException

resource_already_exists_exception

See ResourceAlreadyExistsException

resource_not_found_exception

See ResourceNotFoundException

service_unavailable_exception

See ServiceUnavailableException

unrecognized_client_exception

See UnrecognizedClientException

Structs

AssociateKmsKeyError

CancelExportTaskError

CreateExportTaskError

CreateLogGroupError

CreateLogStreamError

DataAlreadyAcceptedException

The event was already logged.

DeleteDestinationError

DeleteLogGroupError

DeleteLogStreamError

DeleteMetricFilterError

DeleteQueryDefinitionError

DeleteResourcePolicyError

DeleteRetentionPolicyError

DeleteSubscriptionFilterError

DescribeDestinationsError

DescribeExportTasksError

DescribeLogGroupsError

DescribeLogStreamsError

DescribeMetricFiltersError

DescribeQueriesError

DescribeQueryDefinitionsError

DescribeResourcePoliciesError

DescribeSubscriptionFiltersError

DisassociateKmsKeyError

FilterLogEventsError

GetLogEventsError

GetLogGroupFieldsError

GetLogRecordError

GetQueryResultsError

InvalidOperationException

The operation is not valid on the specified resource.

InvalidParameterException

A parameter is specified incorrectly.

InvalidSequenceTokenException

The sequence token is not valid. You can get the correct sequence token in

LimitExceededException

You have reached the maximum number of resources that can be created.

ListTagsLogGroupError

MalformedQueryException

The query string is not valid. Details about this error are displayed in a

OperationAbortedException

Multiple requests to update the same resource were in conflict.

PutDestinationError

PutDestinationPolicyError

PutLogEventsError

PutMetricFilterError

PutQueryDefinitionError

PutResourcePolicyError

PutRetentionPolicyError

PutSubscriptionFilterError

ResourceAlreadyExistsException

The specified resource already exists.

ResourceNotFoundException

The specified resource does not exist.

ServiceUnavailableException

The service cannot complete the request.

StartQueryError

StopQueryError

TagLogGroupError

TestMetricFilterError

UnrecognizedClientException

The most likely cause is an invalid AWS access key ID or secret key.

UntagLogGroupError

Enums

AssociateKmsKeyErrorKind

CancelExportTaskErrorKind

CreateExportTaskErrorKind

CreateLogGroupErrorKind

CreateLogStreamErrorKind

DeleteDestinationErrorKind

DeleteLogGroupErrorKind

DeleteLogStreamErrorKind

DeleteMetricFilterErrorKind

DeleteQueryDefinitionErrorKind

DeleteResourcePolicyErrorKind

DeleteRetentionPolicyErrorKind

DeleteSubscriptionFilterErrorKind

DescribeDestinationsErrorKind

DescribeExportTasksErrorKind

DescribeLogGroupsErrorKind

DescribeLogStreamsErrorKind

DescribeMetricFiltersErrorKind

DescribeQueriesErrorKind

DescribeQueryDefinitionsErrorKind

DescribeResourcePoliciesErrorKind

DescribeSubscriptionFiltersErrorKind

DisassociateKmsKeyErrorKind

FilterLogEventsErrorKind

GetLogEventsErrorKind

GetLogGroupFieldsErrorKind

GetLogRecordErrorKind

GetQueryResultsErrorKind

ListTagsLogGroupErrorKind

PutDestinationErrorKind

PutDestinationPolicyErrorKind

PutLogEventsErrorKind

PutMetricFilterErrorKind

PutQueryDefinitionErrorKind

PutResourcePolicyErrorKind

PutRetentionPolicyErrorKind

PutSubscriptionFilterErrorKind

StartQueryErrorKind

StopQueryErrorKind

TagLogGroupErrorKind

TestMetricFilterErrorKind

UntagLogGroupErrorKind