Module aws_sdk_cloudwatchlogs::model

Modules

destination

See Destination

export_task

See ExportTask

export_task_execution_info

See ExportTaskExecutionInfo

export_task_status

See ExportTaskStatus

filtered_log_event

See FilteredLogEvent

input_log_event

See InputLogEvent

log_group

See LogGroup

log_group_field

See LogGroupField

log_stream

See LogStream

metric_filter

See MetricFilter

metric_filter_match_record

See MetricFilterMatchRecord

metric_transformation

See MetricTransformation

output_log_event

See OutputLogEvent

query_compile_error

See QueryCompileError

query_compile_error_location

See QueryCompileErrorLocation

query_definition

See QueryDefinition

query_info

See QueryInfo

query_statistics

See QueryStatistics

rejected_log_events_info

See RejectedLogEventsInfo

resource_policy

See ResourcePolicy

result_field

See ResultField

searched_log_stream

See SearchedLogStream

subscription_filter

See SubscriptionFilter

Structs

Destination

Represents a cross-account destination that receives subscription log events.

ExportTask

Represents an export task.

ExportTaskExecutionInfo

Represents the status of an export task.

ExportTaskStatus

Represents the status of an export task.

FilteredLogEvent

Represents a matched event.

InputLogEvent

Represents a log event, which is a record of activity that was recorded

LogGroup

Represents a log group.

LogGroupField

The fields contained in log events found by a GetLogGroupFields operation,

LogStream

Represents a log stream, which is a sequence of log events from

MetricFilter

Metric filters express how CloudWatch Logs would extract metric observations

MetricFilterMatchRecord

Represents a matched event.

MetricTransformation

Indicates how to transform ingested log events to metric data in a CloudWatch

OutputLogEvent

Represents a log event.

QueryCompileError

Reserved.

QueryCompileErrorLocation

Reserved.

QueryDefinition

This structure contains details about a saved CloudWatch Logs Insights query definition.

QueryInfo

Information about one CloudWatch Logs Insights query that matches the request in a DescribeQueries operation.

QueryStatistics

Contains the number of log events scanned by the query, the number of log events that matched the

RejectedLogEventsInfo

Represents the rejected events.

ResourcePolicy

A policy enabling one or more entities to put logs to a log group in this account.

ResultField

Contains one field from one log event returned by a CloudWatch Logs Insights query, along with the value of that field.

SearchedLogStream

Represents the search status of a log stream.

SubscriptionFilter

Represents a subscription filter.

Enums

Distribution

The method used to distribute log data to the destination, which can be either

ExportTaskStatusCode

OrderBy

QueryStatus

NOTE: QueryStatus::Unknown has been renamed to ::UnknownValue.

StandardUnit