Module aws_sdk_cloudwatchlogs::operation

Structs

AssociateKmsKey

Associates the specified AWS Key Management Service (AWS KMS) customer master key (CMK) with the specified log group.

CancelExportTask

Cancels the specified export task.

CreateExportTask

Creates an export task, which allows you to efficiently export data from a

CreateLogGroup

Creates a log group with the specified name. You can create up to 20,000 log groups per account.

CreateLogStream

Creates a log stream for the specified log group. A log stream is a sequence of log events

DeleteDestination

Deletes the specified destination, and eventually disables all the

DeleteLogGroup

Deletes the specified log group and permanently deletes all the archived

DeleteLogStream

Deletes the specified log stream and permanently deletes all the archived log events associated

DeleteMetricFilter

Deletes the specified metric filter.

DeleteQueryDefinition

Deletes a saved CloudWatch Logs Insights query definition.

DeleteResourcePolicy

Deletes a resource policy from this account. This revokes

DeleteRetentionPolicy

Deletes the specified retention policy.

DeleteSubscriptionFilter

Deletes the specified subscription filter.

DescribeDestinations

Lists all your destinations. The results are ASCII-sorted by destination name.

DescribeExportTasks

Lists the specified export tasks. You can list all your export tasks or filter

DescribeLogGroups

Lists the specified log groups. You can list all your log groups or filter the results by prefix.

DescribeLogStreams

Lists the log streams for the specified log group.

DescribeMetricFilters

Lists the specified metric filters. You can list all of the metric filters or filter

DescribeQueries

Returns a list of CloudWatch Logs Insights queries that are scheduled, executing, or have

DescribeQueryDefinitions

This operation returns a paginated list of your saved CloudWatch Logs Insights query definitions.

DescribeResourcePolicies

Lists the resource policies in this account.

DescribeSubscriptionFilters

Lists the subscription filters for the specified log group. You can list all the subscription filters or filter the results by prefix.

DisassociateKmsKey

Disassociates the associated AWS Key Management Service (AWS KMS) customer master key (CMK) from the specified log group.

FilterLogEvents

Lists log events from the specified log group. You can list all the log events or filter the results

GetLogEvents

Lists log events from the specified log stream. You can list all of the log events or

GetLogGroupFields

Returns a list of the fields that are included in log events in the specified log group, along with the percentage of log events

GetLogRecord

Retrieves all of the fields and values of a single log event. All fields are retrieved,

GetQueryResults

Returns the results from the specified query.

ListTagsLogGroup

Lists the tags for the specified log group.

PutDestination

Creates or updates a destination. This operation is used only to create destinations for cross-account subscriptions.

PutDestinationPolicy

Creates or updates an access policy associated with an existing

PutLogEvents

Uploads a batch of log events to the specified log stream.

PutMetricFilter

Creates or updates a metric filter and associates it with the specified log group.

PutQueryDefinition

Creates or updates a query definition for CloudWatch Logs Insights. For

PutResourcePolicy

Creates or updates a resource policy allowing other AWS services to put log events to

PutRetentionPolicy

Sets the retention of the specified log group. A retention policy allows you to

PutSubscriptionFilter

Creates or updates a subscription filter and associates it with the specified log

StartQuery

Schedules a query of a log group using CloudWatch Logs Insights. You specify the log group

StopQuery

Stops a CloudWatch Logs Insights query that is in progress. If the query has already ended, the operation

TagLogGroup

Adds or updates the specified tags for the specified log group.

TestMetricFilter

Tests the filter pattern of a metric filter against a sample of log event messages. You

UntagLogGroup

Removes the specified tags from the specified log group.